requirements.txt
Flask
gunicorn
pandas
sklearn
numpy
PyYAML
evidently
dill
matplotlib
seaborn
nbformat
ipykernal
-e .
